Greenium Stable 71.0.0 with User Agent
Greenium 71.0.0 Self-extracting archive.
Just open the exe file, then choose where to extract the files. Then just open
the browser GreeniumStable.exe.
The self-extracting archive is made with 7-zip. This Release includes a shortcut which changes the user agent in the browser, so you can login with a google account.You just have to change the path to the folder where you extracted Greenium.